Second commit
Removed the parallel lists for the id of objects and object themselves and replaced them with a dictionary containing the id's as a key and the user objects as values. Created another dictionary to store books checked out and the date checked out. Created a current time variable and a simulate time option to simulate time going by to implement features like books passed return date and accrued fees on account. Worked a little on the 'View profile information'; however much work is yet to be done. Created a potential log in function to make the code simpler which may or may not be implemented in the future