Third commit
Successfully implemented a simulate time option to simulate weeks going by in the real world. After 2 weeks, a balance will acrue on a user's account and they have the option to view which books need to be returned and by what date. Future notes: need to a create an option to pay off balance,  need to create an option to return a book, need to create smooth transitions after choosing certain options because as of right now the main menu instantly pops up after certain options.